liblab SDK update for version v1.1.72
