+++ examples/install.cmd
@@ -2,4 +2,4 @@ python -m venv .venv
 call .venv\Scripts\activate
 pip install build
 python -m build --outdir dist ..\
-pip install dist\celitech_sdk-1.1.71-py3-none-any.whl --force-reinstall
+pip install dist\celitech_sdk-1.1.72-py3-none-any.whl --force-reinstall
