+++ install.cmd
@@ -15,7 +15,7 @@ if "%USE_VENV%"=="1" (
 
 pip install build
 python -m build --outdir dist .
-pip install dist\celitech_sdk-1.1.71-py3-none-any.whl --force-reinstall
+pip install dist\celitech_sdk-1.1.72-py3-none-any.whl --force-reinstall
 
 if "%USE_VENV%"=="1" (
     deactivate
