+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "celitech-sdk"
-version = "1.1.71"
+version = "1.1.72"
 license = { text = "MIT" }
 description = """Welcome to the CELITECH API documentation!  Useful links: [Homepage](https://www.celitech.com) | [Support email](mailto:support@celitech.com) | [Blog](https://www.celitech.com/blog/) """
 readme = "README.md"
