+++ src/celitech/models/edit_purchase_ok_response.py
@@ -1,5 +1,5 @@
 from .utils.json_map import JsonMap
-from .base import BaseModel
+from .utils.base_model import BaseModel
 
 
 @JsonMap(
