+++ src/celitech/models/get_esim_history_ok_response.py
@@ -1,6 +1,6 @@
 from typing import List
 from .utils.json_map import JsonMap
-from .base import BaseModel
+from .utils.base_model import BaseModel
 
 
 @JsonMap({"status_date": "statusDate", "date_": "date"})
