+++ src/celitech/models/get_purchase_consumption_ok_response.py
@@ -1,5 +1,5 @@
 from .utils.json_map import JsonMap
-from .base import BaseModel
+from .utils.base_model import BaseModel
 
 
 @JsonMap({"data_usage_remaining_in_bytes": "dataUsageRemainingInBytes"})
