+++ src/celitech/net/request_chain/handlers/http_handler.py
@@ -16,12 +16,12 @@ class HttpHandler(BaseHandler):
     :ivar int _timeout_in_seconds: The timeout for the HTTP request in seconds.
     """
 
-    def __init__(self):
+    def __init__(self, timeout=60000):
         """
         Initialize a new instance of HttpHandler.
         """
         super().__init__()
-        self._timeout_in_seconds = 60
+        self._timeout_in_seconds = timeout / 1000
 
     def handle(
         self, request: Request
